+++ gymNotebook.Infrastructure/Commands/Trainings/Training/UpdateTraining.cs
@@ -12,6 +12,6 @@ namespace gymNotebook.Infrastructure.Commands.Trainings.Training
 
         public string Description { get; set; }
 
-        public int Difficulty { get; set; }
+        public byte Difficulty { get; set; }
     }
 }
